stage update

+++ commands/say.js
@@ -1,11 +1,12 @@
 var Discord = require("discord.js");
 var config = require(`../config.json`)
 module.exports.run = async (client, message, args) => {
+  if(!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply("Nie masz uprawnień.");
   message.channel.send(args.join(" "))
   message.delete()
 }
 module.exports.help = {
-	name: "say",
+	name: "admin",
 	category:"util",
   description:"Wysyła wiadomość",
   use:"log!say <wiadomość>"
